requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
opencv-python==4.13.0.92
scikit-learn==1.3.2
scikit-image==0.21.0
matplotlib==3.7.5
seaborn==0.13.2
timm==1.0.25
shap==0.44.1
grad-cam==1.5.5
streamlit==1.40.1
huggingface_hub
